+++ src/actions/actionLogout.ts
