+++ src/actions/actionRecalcProblemScore.ts
@@ -4,8 +4,7 @@ import { PrismaClientKnownRequestError } from "@prisma/client/runtime/library";
 import { cookies } from "next/headers";
 
 import { getUser } from "@/features/auth";
-
-import { recalcProblemScore } from "../queries";
+import { recalcProblemScore } from "@/features/problems";
 
 type ReturnType = { ok: true } | { ok: false; message: string };
