--- src/features/auth/actions/index.ts
@@ -1,3 +1,0 @@
-export * from "./actionAuthenticate";
-export * from "./actionLogin";
-export * from "./actionLogout";
