--- src/features/problems/actions/index.ts
@@ -1,1 +1,0 @@
-export * from "./actionRecalcProblemScore";
