+++ src/features/problems/components/RecalcScore.tsx
@@ -1,9 +1,8 @@
 "use client";
 
+import { actionRecalcProblemScore } from "@/actions";
 import { ResetIcon } from "@/components/icon";
 import { useAppSelector } from "@/hooks";
-
-import { actionRecalcProblemScore } from "../actions";
 
 export const RecalcScore = ({ code }: { code: string }) => {
   const role = useAppSelector((state) => state.user.user?.status);
