--- src/features/submissions/action/index.ts
@@ -1,1 +1,0 @@
-export * from "./actionSubmit";
