+++ src/features/submissions/index.ts
@@ -1,3 +1,2 @@
-export * from "./action";
 export * from "./api";
 export * from "./queries";
