--- src/shared/utils/actions/index.ts
@@ -1,1 +1,0 @@
-export { createProblem } from "./createProblem";
